requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-multipart==0.0.6
pydantic==2.3.0
openai-whisper==20231117
ffmpeg-python==0.2.0 